Moves ColorStyle to a new property package.
